attempted to make better ability button handling, ended up breaking all button handling, rip

+++ ValorantStratPlanner/src/Records/Point.java
@@ -1,8 +1,8 @@
 package Records;
 
 public class Point {
-    public int x, y;
-    public Point(int x, int y){
+    public double x, y;
+    public Point(double x, double y){
         this.x = x;
         this.y = y;
     }
